introduce an optional maven plugin to update the sam config rename module to amazon-lambda-http
