+++ extensions/amazon-lambda-http/maven/src/main/resources/template.yaml
@@ -1,0 +1,19 @@
+---
+AWSTemplateFormatVersion: "2010-09-09"
+Transform: "AWS::Serverless-2016-10-31"
+Globals:
+  Function:
+    Timeout: 5
+Resources:
+  Quarkus:
+    Type: "AWS::Serverless::Function"
+    Properties:
+      Handler: io.quarkus.amazon.lambda.http.HttpHandler::handleRequest+somethingelse
+      Runtime: "java8"
+      Policies: "AWSLambdaBasicExecutionRole"
+      Events:
+        GetResource:
+          Type: "Api"
+          Properties:
+            Path: "/{proxy+}"
+            Method: "any"
