+++ extensions/pom.xml
@@ -105,7 +105,7 @@
 
         <!-- Integrations -->
         <module>amazon-lambda</module>
-        <module>amazon-sam</module>
+        <module>amazon-lambda-http</module>
         <module>amazon-dynamodb</module>
         <module>azure-functions-http</module>
