+++ integration-tests/amazon-lambda-http/src/test/resources/filtered/quarkus.properties
