+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-during-lifecycle/src/main/java/io/quarkus/amazon/sam/maven/SamApp.java
