+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-existing-template/src/main/resources/application.properties
