+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-named-template/src/main/java/io/quarkus/amazon/sam/maven/SamApp.java
