+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-named-template/src/main/resources/application.properties
