+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-new-template/src/main/resources/application.properties
