+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-with-application-properties/properties.yaml
@@ -1,0 +1,29 @@
+Transform: 'AWS::Serverless-2016-10-31'
+Resources:
+
+  properties-resource:
+    # This resource creates a Lambda function.
+    Type: 'AWS::Serverless::Function'
+
+    Properties:
+
+      # This function uses the Nodejs v6.10 runtime.
+      Runtime: nodejs6.10
+
+      # This is the Lambda function's handler.
+      Handler: index.handler
+
+      # The location of the Lambda function code.
+      CodeUri: ./src
+
+      # Event sources to attach to this function. In this case, we are attaching
+      # one API Gateway endpoint to the Lambda function. The function is
+      # called when a HTTP request is made to the API Gateway endpoint.
+      Events:
+
+        ThumbnailApi:
+          # Define an API Gateway endpoint that responds to HTTP GET at /thumbnail
+          Type: Api
+          Properties:
+            Path: /thumbnail
+            Method: GET
