+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-with-application-properties/src/main/java/io/quarkus/amazon/sam/maven/SamApp.java
@@ -1,0 +1,7 @@
+package io.quarkus.amazon.sam.maven;
+
+import javax.ws.rs.Path;
+
+@Path("/sam")
+public class SamApp {
+}
