+++ integration-tests/amazon-lambda-http/src/test/resources/projects/configure-with-application-properties/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+quarkus.amazon-lambda-http.resource-name=properties-resource
+quarkus.amazon-lambda-http.template=properties.yaml
