+++ integration-tests/pom.xml
@@ -58,7 +58,7 @@
         <module>amazon-dynamodb</module>
         <module>test-extension</module>
         <module>amazon-lambda</module>
-        <module>amazon-sam</module>
+        <module>amazon-lambda-http</module>
         <module>kogito</module>
         <module>kogito-maven</module>
         <module>kubernetes-client</module>
